Consolidate the remote tests and cut the comment volume
Tests: 5 artefacts for this feature down to 3, by subject.
  - delete test/bench-strip-null-bytes.mjs — a micro-benchmark with no
    assertions, so it could never fail while still costing a slot in npm test
  - merge test-remote-dedupe-and-doorbell.js and test-remote-heartbeat-tier.js
    into test-remote-transport.js (28 cases: exactly-once, doorbell routing,
    result ordering, heartbeat tiers, reachability/status, capability
    withdrawal, shutdown), which also removes a duplicated Supabase fake
  - test-remote-channel-reconnect.js and test-strip-null-bytes.js keep their
    own subjects

Comments in remote-channel.ts: 498 lines to 308 (36% -> 25% of the file), no
code change. Removed the "an earlier version did X, which broke Y" narratives —
that belongs in git history and is already in the round-6/7/8 commit message —
compressed the constants block, and dropped the degraded-path rationale that was
restated in several places. What stays is the non-obvious mechanics: why the
teardown calls are bounded, why the withdrawal needs 3 attempts and not 2, why
the two heartbeat cadences have to match the server's sweep tiers.

Still above the repo's 12-17% baseline because the new code carries invariants
that are genuinely easy to get wrong, and each surviving note exists because a
review caught it being wrong first.
